feat: Introduce FastAPI application with user authentication, post management, and media upload capabilities.

+++ app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
-from app.db import Post, create_db_and_tables, get_async_session
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession     
 from contextlib import asynccontextmanager  
 from sqlalchemy import select
@@ -19,7 +19,7 @@ async def lifespan(app: FastAPI):
     
 app = FastAPI(lifespan = lifespan)
 
-app.include_router(fastapi_users.get_auth_router(auth_backend),prefix="/auth",tags=["auth"])
+app.include_router(fastapi_users.get_auth_router(auth_backend),prefix="/auth/jwt",tags=["auth"])
 app.include_router(fastapi_users.get_register_router(UserRead,UserCreate),prefix="/auth",tags=["auth"])
 app.include_router(fastapi_users.get_reset_password_router(),prefix="/auth",tags=["auth"])
 app.include_router(fastapi_users.get_verify_router(UserRead),prefix="/auth",tags=["auth"])
@@ -29,6 +29,7 @@ app.include_router(fastapi_users.get_users_router(UserRead,UserUpdate),prefix="/
 async def upload_file(
         file: UploadFile = File(...),
         caption: str = Form(...),
+        user: User = Depends(current_active_user),
         session: AsyncSession = Depends(get_async_session)
 ):
     temp_file_path = None
@@ -50,11 +51,11 @@ async def upload_file(
             raise HTTPException(status_code=500, detail="Image upload failed")
 
         post = Post(
+            user_id = user.id,
             caption = caption,
             url = upload_result.url,
             file_type = "video" if file.content_type.startswith("video/") else "image",
-            file_name = upload_result.name,
-            user_id = None
+            file_name = upload_result.name
         )
         session.add(post)
         await session.commit()
@@ -76,33 +77,47 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-        session: AsyncSession = Depends(get_async_session)
+        session: AsyncSession = Depends(get_async_session),
+        user: User = Depends(current_active_user)
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = result.scalars().all()
 
-    post_data =[]
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    users_dict = {u.id: u.email for u in users}
+
+    post_data = []
     for post in posts:
         post_data.append(PostResponse(
-            id = post.id,
-            caption = post.caption,
-            url = post.url,
-            file_type = post.file_type,
-            file_name = post.file_name, 
-            created_at = post.created_at.isoformat()
+            id=post.id,
+            user_id=str(post.user_id),
+            caption=post.caption,
+            url=post.url,
+            file_type=post.file_type,
+            file_name=post.file_name,
+            created_at=post.created_at.isoformat(),
+            is_owner=post.user_id == user.id,
+            email=users_dict.get(post.user_id, "Unknown")
         ))
     return {"posts": post_data}
 
 @app.delete("/post/{post_id}")
 async def delete_post(
         post_id: uuid.UUID,
-        session: AsyncSession = Depends(get_async_session)
+        session: AsyncSession = Depends(get_async_session),
+        user: User = Depends(current_active_user)
 ):
     try:
         result = await session.execute(select(Post).where(Post.id == post_id))
         post = result.scalars().first()
+
+
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
+
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You are not authorized to delete this post")
         await session.delete(post)
         await session.commit()
         return {"success": True, "message": "Post deleted successfully"}
